# cellp/api/openapi.yaml
openapi: 3.0.3
info:
  title: cellp API
  version: 1.0.0
  description: cellp control plane REST API (DESIGN §9)
servers:
  - url: http://127.0.0.1:8790/v1
paths:
...
  /projects/{projectID}/versions/{versionID}:
    get:
      summary: Get version status
      operationId: getVersion
      security:
        - adminToken: []
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
        - name: versionID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    delete:
      summary: Destroy version
      operationId: destroyVersion
      security:
        - adminToken: []
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
        - name: versionID
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Accepted (draining)
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: draining
                  id:
                    type: string
                  project_id:
                    type: string
...
components:
...
  securitySchemes:
    deployToken:
      type: http
      scheme: bearer
      description: CELLP_DEPLOY_TOKEN or PLATFORM_TOKEN — POST /versions only
    adminToken:
      type: http
      scheme: bearer
      description: CELLP_ADMIN_TOKEN or PLATFORM_TOKEN (dev-local-token fallback) — full admin